+++ src/main/java/com/alper/worldcup/service/UserMatchStats.java
@@ -1,3 +1,3 @@
 package com.alper.worldcup.service;
 
-public record UserMatchStats(int exactScores, int correctOutcomes) {}
+public record UserMatchStats(int exactScores, int correctOutcomes, int missedGames) {}